improve code sharing in SymCC fuzz targets
