+++ cedar-drt/fuzz/fuzz_targets/symcc-smt-script-drt.rs
@@ -16,24 +16,17 @@
 
 #![no_main]
 use cedar_drt::logger::initialize_log;
-use cedar_drt_inner::{fuzz_target, symcc::total_action_request_env_limit};
+use cedar_drt_inner::{fuzz_target, symcc::SinglePolicyFuzzTargetInput};
 
 use cedar_lean_ffi::CedarLeanFfi;
 use cedar_policy::{Policy, PolicySet, Schema};
-use cedar_policy_generators::{
-    abac::ABACPolicy, hierarchy::HierarchyGenerator, schema, schema_gen::SchemaGen,
-    settings::ABACSettings,
-};
-
-use libfuzzer_sys::arbitrary::{self, Arbitrary, MaxRecursionReached, Unstructured};
-use log::debug;
-use std::convert::TryFrom;
-
 use cedar_policy_symcc::{
     always_allows_asserts, solver::WriterSolver, CedarSymCompiler, CompiledPolicySet,
     WellFormedAsserts,
 };
 
+use log::debug;
+use std::convert::TryFrom;
 use std::sync::LazyLock;
 
 static RUNTIME: LazyLock<tokio::runtime::Runtime> = LazyLock::new(|| {
@@ -42,42 +35,6 @@ static RUNTIME: LazyLock<tokio::runtime::Runtime> = LazyLock::new(|| {
         .build()
         .unwrap()
 });
-
-/// Input expected by this fuzz target
-#[derive(Debug, Clone)]
-pub struct FuzzTargetInput {
-    /// generated schema
-    pub schema: schema::Schema,
-    /// generated policy
-    pub policy: ABACPolicy,
-}
-
-/// settings for this fuzz target
-const SETTINGS: ABACSettings = ABACSettings {
-    max_depth: 3,
-    max_width: 3,
-    total_action_request_env_limit: total_action_request_env_limit(),
-    ..ABACSettings::type_directed()
-};
-
-impl<'a> Arbitrary<'a> for FuzzTargetInput {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let schema = schema::Schema::arbitrary(SETTINGS.clone(), u)?;
-        let hierarchy = schema.arbitrary_hierarchy(u)?;
-        let policy = schema.arbitrary_policy(&hierarchy, u)?;
-
-        Ok(Self { schema, policy })
-    }
-
-    fn try_size_hint(
-        depth: usize,
-    ) -> std::result::Result<(usize, Option<usize>), MaxRecursionReached> {
-        Ok(arbitrary::size_hint::and_all(&[
-            schema::Schema::arbitrary_size_hint(depth)?,
-            HierarchyGenerator::size_hint(depth),
-        ]))
-    }
-}
 
 fn smtlib_of_check_asserts(rust_asserts: &WellFormedAsserts<'_>) -> Result<String, String> {
     RUNTIME.block_on(async {
@@ -96,7 +53,7 @@ fn smtlib_of_check_asserts(rust_asserts: &WellFormedAsserts<'_>) -> Result<Strin
 
 // Fuzzing target checking that Term to SMT-LIB encodings generated by Lean and Rust SymCC implementations are equivalent
 // cspell:ignore symcc
-fuzz_target!(|input: FuzzTargetInput| {
+fuzz_target!(|input: SinglePolicyFuzzTargetInput| {
     initialize_log();
     let lean_ffi = CedarLeanFfi::new();
     let mut policyset = PolicySet::new();
