+++ cedar-drt/fuzz/fuzz_targets/symcc-term-roundtrip-abac-type-directed.rs
@@ -17,61 +17,17 @@
 #![no_main]
 use cedar_drt::logger::initialize_log;
 
-use cedar_drt_inner::{fuzz_target, symcc::total_action_request_env_limit};
+use cedar_drt_inner::{fuzz_target, symcc::SinglePolicyFuzzTargetInput};
 
 use cedar_lean_ffi::CedarLeanFfi;
 use cedar_policy::{Policy, PolicySet, Schema};
 
-use cedar_policy_generators::{
-    abac::ABACPolicy, hierarchy::HierarchyGenerator, schema, schema_gen::SchemaGen,
-    settings::ABACSettings,
-};
-
-use libfuzzer_sys::arbitrary::{self, Arbitrary, MaxRecursionReached, Unstructured};
 use log::debug;
 use similar_asserts::assert_eq;
 use std::convert::TryFrom;
 
-/// Input expected by this fuzz target:
-/// An ABAC hierarchy, policy, and 8 associated requests
-#[derive(Debug, Clone)]
-pub struct FuzzTargetInput {
-    /// generated schema
-    pub schema: schema::Schema,
-    /// generated policy
-    pub policy: ABACPolicy,
-}
-
-/// settings for this fuzz target
-const SETTINGS: ABACSettings = ABACSettings {
-    max_depth: 3,
-    max_width: 3,
-    total_action_request_env_limit: total_action_request_env_limit(),
-    ..ABACSettings::type_directed()
-};
-
-impl<'a> Arbitrary<'a> for FuzzTargetInput {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let schema = schema::Schema::arbitrary(SETTINGS.clone(), u)?;
-        let hierarchy = schema.arbitrary_hierarchy(u)?;
-        let policy = schema.arbitrary_policy(&hierarchy, u)?;
-
-        Ok(Self { schema, policy })
-    }
-
-    fn try_size_hint(
-        depth: usize,
-    ) -> std::result::Result<(usize, Option<usize>), MaxRecursionReached> {
-        Ok(arbitrary::size_hint::and_all(&[
-            schema::Schema::arbitrary_size_hint(depth)?,
-            HierarchyGenerator::size_hint(depth),
-            schema::Schema::arbitrary_policy_size_hint(&SETTINGS, depth),
-        ]))
-    }
-}
-
 // Fuzzing Target to show that Asserts/Term Serialization/Deserialization does not effect the final SMTLib script produced
-fuzz_target!(|input: FuzzTargetInput| {
+fuzz_target!(|input: SinglePolicyFuzzTargetInput| {
     initialize_log();
     let lean_ffi = CedarLeanFfi::new();
     let mut policyset = PolicySet::new();
